+++ src/udp/mod.rs
@@ -15,6 +15,7 @@ pub mod errors;
 pub mod request;
 pub mod server;
 pub mod handlers;
+pub mod time_bound_pepper;
 
 pub type Bytes = u64;
 pub type Port = u16;
